moved files to correct directories, updated includes to demonstrate this
